Reworked some classes

+++ src/main/kotlin/dev/themeinerlp/skinserver/config/HeadView.kt
@@ -1,11 +1,11 @@
 package dev.themeinerlp.skinserver.config
 
-enum class HeadView(val leftHead: Float, val topHead: Float, val overlayLeft: Float, val overlayTop: Float) {
-    Front(8F, 8F, 40F, 8F),
-    Right(0F, 8F, 32F, 8F),
-    Left(16F, 8F, 48F, 8F),
-    Back(24F, 8F, 56F, 8F),
-    Top(8F, 0F, 40F, 0F),
-    Bottom(16F, 0F, 48F, 0F),
+enum class HeadView(val leftHead: Int, val topHead: Int, val overlayLeft: Int, val overlayTop: Int) {
+    Front(8, 8, 40, 8),
+    Right(0, 8, 32, 8),
+    Left(16, 8, 48, 8),
+    Back(24, 8, 56, 8),
+    Top(8, 0, 40, 0),
+    Bottom(16, 0, 48, 0),
 
 }
